+++ lib/boruta/oauth/schemas/client.ex
@@ -195,21 +195,30 @@ defmodule Boruta.Oauth.Client do
 
   @spec should_check_secret?(client :: t(), grant_type :: String.t()) :: boolean()
   def should_check_secret?(_client, grant_type)
-      when grant_type in ["implicit", "code", "preauthorized_code"],
+      when grant_type in ["implicit", "code"],
       do: false
 
   def should_check_secret?(client, grant_type) when grant_type in ["refresh_token", "revoke"] do
     not apply(__MODULE__, :"public_#{grant_type}?", [client])
   end
 
-  def should_check_secret?(%__MODULE__{confidential: true}, _grant_type), do: true
-
   def should_check_secret?(_client, grant_type)
       when grant_type in ["client_credentials", "agent_credentials", "introspect"],
       do: true
 
-  def should_check_secret?(%__MODULE__{public_client_id: "" <> client_id}, _grant_type),
-    do: client_id != issuer()
+  def should_check_secret?(
+        %__MODULE__{public_client_id: "" <> _client_id, confidential: true},
+        _grant_type
+      ),
+      do: true
+
+  def should_check_secret?(
+        %__MODULE__{public_client_id: "" <> client_id, confidential: false},
+        _grant_type
+      ),
+      do: client_id != issuer()
+
+  def should_check_secret?(%__MODULE__{confidential: true}, _grant_type), do: true
 
   def should_check_secret?(%__MODULE__{confidential: false}, _grant_type), do: false